requirements.txt
click==7.1.2
cycler==0.10.0
defusedxml==0.6.0
fastapi==0.65.1
h11==0.12.0
h5py==2.10.0
isodate==0.6.0
joblib==1.0.1
Keras==2.4.3
keras-pickle-wrapper==1.0.5
kiwisolver==1.3.1
lxml==4.6.3
matplotlib==3.4.2
mpmath==1.2.1
ndindex==1.5.2
numpy==1.18.1
opencv-python==4.5.2.52
p5py==1.0.0
pandas==1.2.4
pep517==0.10.0
Pillow==7.0.0
pydantic==1.8.2
pyparsing==2.4.7
python-dateutil==2.8.1
python-multipart==0.0.5
python3-saml==1.10.1
pytz==2021.1
PyYAML==5.4.1
scikit-learn==0.24.2
scipy==1.6.3
six==1.16.0
starlette==0.14.2
sympy==1.8
threadpoolctl==2.1.0
toml==0.10.2
typing-extensions==3.10.0.0
uvicorn==0.13.4
versioned-hdf5==1.3
xmlsec==1.3.10
